+++ src/main/java/com/github/aytchell/qrgen/renderers/ImgParameters.java
@@ -10,4 +10,5 @@ class ImgParameters {
     int firstCellY;
     int onColor;
     int offColor;
+    int markerColor;
 }